refractored return types in route
created extra class s return type for some methods in routes
+ added first few tests for bot moving
