+++ app/src/main/java/com/example/scotlandyard/map/motions/MarkerAnimation.java
@@ -111,72 +111,3 @@ public class MarkerAnimation {
         handler.post(motions.get(0));
     }
 }
-
-class MarkerMotion implements Runnable {
-
-    private Marker marker;
-    private LatLng nextPoint;
-    private LatLngInterpolator latLngInterpolator;
-    private float duration;
-    private long elapsed;
-    private long start;
-    private float t;
-    private Interpolator interpolator;
-    private MarkerMotion nextMotion;
-    private Handler handler;
-    private LatLng current;
-    private boolean showToast;
-    private Context toastContext;
-    private int finalIcon;
-
-    public MarkerMotion(Marker marker, LatLng nextPoint, LatLngInterpolator latLngInterpolator, float duration, int finalIcon) {
-        this.marker = marker;
-        this.nextPoint = nextPoint;
-        this.latLngInterpolator = latLngInterpolator;
-        this.duration = duration;
-        this.elapsed = 0l;
-        this.t = 0f;
-        this.interpolator = new AccelerateDecelerateInterpolator();
-        this.nextMotion = null;
-        this.current = marker.getPosition();
-        this.start = SystemClock.uptimeMillis();
-        this.finalIcon = finalIcon;
-    }
-
-    public void setNextMotion(MarkerMotion nextMotion) {
-        this.nextMotion = nextMotion;
-    }
-
-    public void setHandler(Handler handler) {
-        this.handler = handler;
-    }
-
-    public void setMarker(Marker marker) {
-        this.marker = marker;
-        this.current = marker.getPosition();
-    }
-
-    public void setStart() {
-        this.start = SystemClock.uptimeMillis();
-    }
-
-    @Override
-    public void run() {
-        elapsed = SystemClock.uptimeMillis() - start;
-        t = elapsed / duration;
-        float v = interpolator.getInterpolation(t);
-        marker.setPosition(latLngInterpolator.interpolate(v, current, nextPoint));
-        if (t < 1.0) {
-            handler.postDelayed(this, 16);
-        } else {
-            marker.setPosition(nextPoint);
-            if (nextMotion != null) {
-                nextMotion.setMarker(marker);
-                nextMotion.setStart();
-                handler.post(nextMotion);
-            } else {
-                marker.setIcon(BitmapDescriptorFactory.fromResource(finalIcon));
-            }
-        }
-    }
-}
