+++ app/src/main/java/com/example/scotlandyard/map/routefactories/FootRoutes.java
@@ -1,0 +1,245 @@
+package com.example.scotlandyard.map.routefactories;
+
+import com.example.scotlandyard.map.Point;
+import com.example.scotlandyard.map.Route;
+
+public class FootRoutes {
+
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_5_6 = {
+            new Point(46.618745, 14.261462),
+            new Point(46.618712, 14.261580),
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_20_21 = {
+            new Point(46.620758, 14.264536),
+            new Point(46.620544, 14.264391)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_21_22 = {
+            new Point(46.620322, 14.264098),
+            new Point(46.620281, 14.263773),
+            new Point(46.620164, 14.263646)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_28_29 = {
+            new Point(46.617871, 14.267544),
+            new Point(46.617710, 14.267488)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_21_23 = {
+            new Point(46.620322, 14.264098),
+            new Point(46.620149, 14.264186),
+            new Point(46.620059, 14.264390),
+            new Point(46.619938, 14.264483)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_21_24 = {
+            new Point(46.620322, 14.264098),
+            new Point(46.620252, 14.264625),
+            new Point(46.619966, 14.265080)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_22_23 = {
+            new Point(46.619799, 14.263644),
+            new Point(46.619712, 14.263757),
+            new Point(46.619694, 14.264272),
+            new Point(46.619821, 14.264491)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_23_24 = {
+            new Point(46.620018, 14.264658),
+            new Point(46.619990, 14.264842),
+            new Point(46.619902, 14.264867)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_24_27 = {
+            new Point(46.619743, 14.265144),
+            new Point(46.619589, 14.265078),
+            new Point(46.619292, 14.265090),
+            new Point(46.619268, 14.265285)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_24_26 = {
+            new Point(46.619624, 14.264825),
+            new Point(46.619516, 14.264847),
+            new Point(46.619496, 14.264725),
+            new Point(46.619354, 14.264780),
+            new Point(46.619262, 14.264750)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_26_27 = {
+            new Point(46.619262, 14.264750),
+            new Point(46.619233, 14.264938),
+            new Point(46.619292, 14.265090),
+            new Point(46.619268, 14.265285)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_22_25 = {
+            new Point(46.619693, 14.263541),
+            new Point(46.619650, 14.263599),
+            new Point(46.619327, 14.263563),
+            new Point(46.619357, 14.263670)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_25_26 = {
+            new Point(46.619375, 14.264148)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_10_38 = {
+            new Point(46.619035, 14.266970)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_28_35 = {
+            new Point(46.617947, 14.268011),
+            new Point(46.617894, 14.268020),
+            new Point(46.617891, 14.268264),
+            new Point(46.617975, 14.268311)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_29_35 = {
+            new Point(46.617451, 14.267580),
+            new Point(46.617546, 14.267736),
+            new Point(46.617672, 14.267704),
+            new Point(46.617681, 14.268023),
+            new Point(46.617975, 14.268311)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_29_36 = {
+            new Point(46.617385, 14.267708),
+            new Point(46.617206, 14.267824),
+            new Point(46.617210, 14.268709)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_36_37 = {
+            new Point(46.617110, 14.268567),
+            new Point(46.616641, 14.268528)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_42_51 = {
+            new Point(46.617284, 14.263694),
+            new Point(46.617202, 14.262918)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_45_46 = {
+            new Point(46.617133, 14.267060)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_46_47 = {
+            new Point(46.616966, 14.267022),
+            new Point(46.617067, 14.266893)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_46_53 = {
+            new Point(46.616565, 14.267152)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_47_48 = {
+            new Point(46.617016, 14.266255)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_47_54 = {
+            new Point(46.616724, 14.266393)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_50_51 = {
+            new Point(46.616897, 14.263025),
+            new Point(46.617061, 14.263001)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_53_61 = {
+            new Point(46.616351, 14.266395)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_56_62 = {
+            new Point(46.616211, 14.265390),
+            new Point(46.615729, 14.265454)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_59_73 = {
+            new Point(46.615511, 14.261994)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_62_63 = {
+            new Point(46.615581, 14.265248)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_64_65 = {
+            new Point(46.615629, 14.263624),
+            new Point(46.615572, 14.263472),
+            new Point(46.615478, 14.263423)
+    };
+    private static final Point[] INTERMEDIATES_BY_FOOT_PTS_71_74 = {
+            new Point(46.613368, 14.264279)
+    };
+
+    public static final Route[] BY_FOOT = {
+            new Route(1, 2),
+            new Route(2, 3),
+            new Route(2, 4),
+            new Route(3, 5),
+            new Route(4, 6),
+            new Route(5, 6, INTERMEDIATES_BY_FOOT_PTS_5_6),
+            new Route(6, 7),
+            new Route(7, 8),
+            new Route(8, 9),
+            new Route(9, 10),
+            new Route(10, 11),
+            new Route(10, 18),
+            new Route(10, 38, INTERMEDIATES_BY_FOOT_PTS_10_38),
+            new Route(12, 13),
+            new Route(13, 14),
+            new Route(14, 15),
+            new Route(16, 17),
+            new Route(17, 24),
+            new Route(19, 20),
+            new Route(20, 21, INTERMEDIATES_BY_FOOT_PTS_20_21),
+            new Route(21, 24, INTERMEDIATES_BY_FOOT_PTS_21_24),
+            new Route(21, 23, INTERMEDIATES_BY_FOOT_PTS_21_23),
+            new Route(21, 22, INTERMEDIATES_BY_FOOT_PTS_21_22),
+            new Route(22, 23, INTERMEDIATES_BY_FOOT_PTS_22_23),
+            new Route(22, 25, INTERMEDIATES_BY_FOOT_PTS_22_25),
+            new Route(23, 24, INTERMEDIATES_BY_FOOT_PTS_23_24),
+            new Route(24, 27, INTERMEDIATES_BY_FOOT_PTS_24_27),
+            new Route(24, 26, INTERMEDIATES_BY_FOOT_PTS_24_26),
+            new Route(25, 26, INTERMEDIATES_BY_FOOT_PTS_25_26),
+            new Route(26, 27, INTERMEDIATES_BY_FOOT_PTS_26_27),
+            new Route(28, 38),
+            new Route(28, 35, INTERMEDIATES_BY_FOOT_PTS_28_35),
+            new Route(28, 29, INTERMEDIATES_BY_FOOT_PTS_28_29),
+            new Route(28, 39),
+            new Route(29, 36, INTERMEDIATES_BY_FOOT_PTS_29_36),
+            new Route(29, 35, INTERMEDIATES_BY_FOOT_PTS_29_35),
+            new Route(29, 45),
+            new Route(30, 37),
+            new Route(30, 53),
+            new Route(30, 60),
+            new Route(31, 66),
+            new Route(32, 66),
+            new Route(33, 36),
+            new Route(33, 37),
+            new Route(34, 35),
+            new Route(36, 37, INTERMEDIATES_BY_FOOT_PTS_36_37),
+            new Route(38, 39),
+            new Route(40, 41),
+            new Route(41, 48),
+            new Route(42, 49),
+            new Route(42, 50),
+            new Route(42, 51, INTERMEDIATES_BY_FOOT_PTS_42_51),
+            new Route(43, 52),
+            new Route(43, 44),
+            new Route(43, 51),
+            new Route(45, 46, INTERMEDIATES_BY_FOOT_PTS_45_46),
+            new Route(46, 47, INTERMEDIATES_BY_FOOT_PTS_46_47),
+            new Route(46, 53, INTERMEDIATES_BY_FOOT_PTS_46_53),
+            new Route(47, 48, INTERMEDIATES_BY_FOOT_PTS_47_48),
+            new Route(47, 54, INTERMEDIATES_BY_FOOT_PTS_47_54),
+            new Route(48, 49),
+            new Route(50, 51, INTERMEDIATES_BY_FOOT_PTS_50_51),
+            new Route(51, 52),
+            new Route(52, 58),
+            new Route(53, 61, INTERMEDIATES_BY_FOOT_PTS_53_61),
+            new Route(54, 55),
+            new Route(55, 61),
+            new Route(55, 56),
+            new Route(56, 62, INTERMEDIATES_BY_FOOT_PTS_56_62),
+            new Route(56, 63),
+            new Route(56, 75),
+            new Route(57, 64),
+            new Route(57, 75),
+            new Route(57, 58),
+            new Route(58, 59),
+            new Route(59, 73, INTERMEDIATES_BY_FOOT_PTS_59_73),
+            new Route(60, 61),
+            new Route(60, 66),
+            new Route(66, 67),
+            new Route(62, 63, INTERMEDIATES_BY_FOOT_PTS_62_63),
+            new Route(63, 64),
+            new Route(63, 68),
+            new Route(64, 65, INTERMEDIATES_BY_FOOT_PTS_64_65),
+            new Route(65, 69),
+            new Route(65, 72),
+            new Route(68, 67),
+            new Route(68, 69),
+            new Route(69, 71),
+            new Route(70, 71),
+            new Route(71, 72),
+            new Route(71, 74, INTERMEDIATES_BY_FOOT_PTS_71_74),
+            new Route(72, 74),
+            new Route(72, 73)
+    };
+
+    private FootRoutes() {
+    }
+
+}
